fix: Remove Node.js-specific security modules incompatible with CEP environment
- Remove SecureCommandExecutor and ValidationService that use Node.js modules (child_process, os)
- Update openTASFolder to use CEP Node interface directly for cross-platform folder opening
- Fix TypeScript error handling in openTASFolder
- Update SecureSocketManager to use standard JSON.parse instead of ValidationService
- Fix processingStore enum imports (import as values, not types)

These modules were causing build errors as they attempted to use Node.js-specific
modules in the browser context of the CEP panel. CEP extensions must access Node.js
functionality through the CEP Node interface, not directly.
